[Mujoco parser] Silently ignore some attributes/elements.
Now that MuJoCo has space in `parsing_doxygen.h` where we can list
"silently ignored" attributes and elements, this PR goes through and
silences those elements that we will never support (and reasonable
users would not expected us to support).

This will help cut down a bit on the substantial noise coming out of
the parser.

It also adds a few new warnings for new attributes and elements that
have managed to spring up in the mujoco documentation.