Correct parsing for multiple materials
If a material specifies a diffuse map (map_Kd) but doesn't specify a
Kd value, it *should* default to white. However, for multiple materials
in the file, the parser makes an error. It attempts to track whether the
material has specified Kd. But it never resets that record. So, as soon as
any material specifies Kd, all subsequent material definitions are
classified as *also* having provided one -- even if they haven't. This
leads to the correctly reset diffuse color of black being used with
diffuse textures (producing black meshes).

This simply resets the `has_kd` flag every time a new material definition
starts.