Modify install_prereqs to create a venv (#22477)
Modify the macOS install_prereqs script used for Drake tarball
installations to create a Python virtual environment in the location
where Drake has been extracted. Also use this to install Drake's Python
dependencies using PDM. Update user documentation accordingly.

This means we are now using the pinned versions of our dependencies and
are no longer keeping a duplicate copy of this list solely for
installations.

This also means that we are installing a pinned numpy in the venv for
installations. This was already the case for source builds, but means we
now no longer have any reason to install numpy from Homebrew, so remove
it from there.