[py multibody] Adjust element-adding API to avoid unique_ptr
This reduces our reliance RLG/pybind11 custom unique_ptr semantics.

Introduce a new DoShallowClone virtual function on Joint, Frame, and
ForceElement. Downstream code that has implemented its own subclass of
any of those bases and bound the subclass in Python must override the
new virtual function.

Python code must no longer assume that the object passed AddJoint,
AddFrame, or AddForceElement is what gets added to the plant. Instead,
code should always capture the return value, i.e.,

  joint = plant.AddJoint(PrismaticJoint(...))

or

  joint = PrismaticJoint(...)
  joint = plant.AddJoint(joint)

are fine, but in this approach

  joint = PrismaticJoint(...)
  plant.AddJoint(joint)

the original 'joint' object will NOT necessarily be part of the plant,
so would probably be a mistake to be used thereafter.