[build] Optionally build zlib from source
Switch the wheel build to use the source build. Even though zlib is
part of the required set of dynamic libraries wheels are allowed to
universally rely on, it's easier to just remain consistent and set all
of the options the same way.

This is the last of the tools/wheel/image/dependencies, so we also
remove all of that infrastructure as well.