Desired state input ports are no longer required to be connected.
This PR changes the requirements on desired input ports in the following
ways:
 1. Partially PD-controlled model instances are allowed. An exception is
    no longer thrown when evaluating desired state inputs for partically
    controlled model instances.
 2. This port is no longer required to be connected. If disconnected,
    the corresponding model instance's controllers are marked as
    'armed', and have no effect on the dynamics of the plant.
 3. Desired state values corresponding to actuators with no PD control
    are ignored. In particular, even NaNs are ignored, no longer
    causing an exception.

Before this PR, evaluation of an input port for a partially
PD-controlled model instance would trigger an exception.
Now

If disconnected, model-instance desired-state input ports lead to
disarmed PD controllers for that model. Therefore they are no longer
requied to be connected.