Fix CMake dependencies
Add an explicit dependency on the drake_version target to the
drake_cxx_python target. Previously, the latter depended on the
generated version script, but that isn't understood as an output of the
drake_version target, which would sometimes lead to build errors due to
the file not existing and the tool not knowing how to build it. (This is
more likely to be a problem with build parallelism enabled, and
especially was hitting Ninja builds.)