Switch non-test code to use Parser(builder) (#22662)
PR 22531 significantly expanded the role of the Parser from just
parsing into MultibodyPlant to (potentially) also parsing into a
DiagramBuilder. Because it is strictly more general, calling
`Parser(builder)` instead of `Parser(plant)` is now the preferred
usage when the standard builder/plant combination are both
available. This PR updates non-test (library and example) code to use
it, in order to set a good example.

Resolves #22641.