Extend Mesh/Convex to include non-uniform scale
Note: while this adds the APIs, the non-uniform scale is not used in
most places. Configuring a non-uniform scale and then exercising those
domains will throw an exception.

Currently, the computation of convex hulls from non-uniform scale is the
only supported operation.

NOTE TO RELEASE ENGINEER:

This is one of multiple PRs that enable non-uniform scaling (see also
#22772).  For the release, they should all be lumped together. The total
feature description is:

Enable non-uniform scaling for Mesh and Convex geometries.