Build wheels on AlmaLinux instead of Ubuntu
Change wheel builds to use AlmaLinux instead of Ubuntu as the host
platform. The motivating effect of this is to "downgrades" our required
glibc version from 2.35 to 2.34.

Note that AlmaLinux 9 is the base of manylinux_2_34_x86_64:2025.01.05-1.
The reason we are not using manylinux is because a) it provides Python
interpreters that do not have headers or libraries, and so are not
useful for building Drake, and b) its compilers interfere with use of
the underlying AlmaLinux compilers and attempts to use them result in
obscure failures.