Compare RenderEngineParams with existing RenderEngine instances (#22871)
This provides the ability to ask a RenderEngine instance if a set of
parameters describes the engine.

We apply this new ability when processing CameraConfig instances. It makes
the process more forgiving as multiple CameraConfig can reference
otherwise equal RenderEngine instances (whereas before, it was more
brittle and inclined to throw).