Add explicit requirement for diffutils
Our patch-applying process somewhere appears to rely on `diff` (normally
provided on Linux via the GNU diffutils package). Previously, we have
taken it for granted that this was available, but AlmaLinux appears to
have "trimmed" their Docker image recently to omit it, resulting in
build failures. Add the providing package to our various dependency
lists to ensure it is available.