[fem] Parallel residual and tangent matrix computation
Push more per-element computation into the parallel ComputeData
function. This includes the inverser dynamics force, stiffness matrix,
damping matrix, and the tangent matrix.

Remove the corresponding Calc methods from the FemElement class so that
derived elements use their own signature to compute these quantities
that better indicates the dependencies and allows more opportunistic
optimizations.

The integrator weights are added to the FemState class as a parameter so
that they could be passed to the element computation.