Factor Out Common Options Between IRIS-ZO and IRIS-NP2 (#22982)
This moves options that are intended to be common to both IRIS-ZO and IRIS-NP2 into a single class, CommonIrisOptions. It also separates out the parameterization functionality.

We also update the default value of max_separating_planes_per_iteration per current best practices, and reorganize the constructors and methods of IrisParameterizationFunction for better consistency.

Finally, we include two fixes:
* Fix an issue with how the "note" block regarding GIL is rendered in the pydrake documentation.
* Fix a bindings bug where an InverseKinematics object may be deleted while its subsidiary MathematicalProgram is still being used, leading to a segmentation fault.