GetBodiesKinematicallyAffectedBy no longer throws on welds
Previously, it would throw if the index of a weld joint is provided.
The underlying implementation already did the "right" thing in the face of
a weld joint -- nothing. So, rather than throwing, we simply clarify
the documentation.

This isn't breaking, because all code that currently works will continue
to work. The function just becomes more relaxed.