Guard some Eigen realloc calls against zero (#23056)
Calling 'realloc(size=0)' is typically a synonym for 'free', but this
is not portable, and newer Valgrind has started flagging this.

Eigen 3.4.0 assumes the non-portable free during conservativeResize,
so we need to clean up some our call sites that tickle that case.

This has been fixed upstream but not released yet:
https://gitlab.com/libeigen/eigen/-/merge_requests/1148