[tools] Fix push_release for the source code archive
Amend bbcde5ab0 after testing on v1.42.0.

* Open the file being uploaded to the release. (The previous revision
simply uploaded the file name as a string, not the file itself.)
* Hard-code the content type of the downloaded files, since `mimetypes`
fails on both the tarballs and shas, and no other package which
implements this functionality is available in the standard library.
* Add `repo` permissions to GitHub tokens so that files can be pushed.

Also add return type annotations to all functions in push_release.py.