Fix macOS wheel RECORD (#23294)
Rewrite logic to remove the MOSEK libraries from the macOS wheel to use
the 'wheel' command to unpack and repack the wheel in order to perform
modifications, rather than trying to perform the modifications in a more
'in place' manner. This is needed so that the wheel's RECORD is updated
to reflect the changes.

Accordingly, remove logic in 'change_lpath' to edit files inside wheels,
since doing so results in a broken RECORD.