Build wheels with 'build'
Update the wheel build process to use the 'build' package to build the
wheels instead of invoking 'setup.py' directly (which is deprecated).

Also, mark 'wheel' as only needed on macOS. It used to be needed for the
'bdist_wheel' command of 'setuptools', but that hasn't been the case for
about a year. However, we recently started using it to alter the macOS
wheel (to remove MOSEK), so it isn't dropped entirely.