[tools] Retrieve GitHub source archive manually (#23309)
The GitHub API endpoint that `github3` uses to fetch the source archive
(downloaded and reuploaded with checksums to both GitHub and S3) ends up
with `RobotLocomotion-drake-<commit hash>` as the root folder inside the
tarball instead of `drake-<version>`. The latter mirrors the download
link provided on the release, so it's what this script should use.

Use `urllib` to retrieve it because `github3` does not provide a way of
accessing the endpoint needed.