[build] Optionally install a static `libdrake.a`
Build static and shared copies of `libdrake`, and provide CMake and
Bazel options to decide which one to install.

* Add corresponding CMake and Bazel options for whether to install
Python tools, which is incompatible with installing the static library.
Additionally, it's incompatible with the LCM runtime.
* Rename the `drake_cxx_python` CMake target to `drake_install`, since
the "Python" part is now conditional.
* Mark the installed `drake::drake-marker` CMake target for deprecation,
since it's intended for internal use only.
* Vendor Bazel's currently-experimental `cc_static_library` to build the
library.

For now, `libdrake.a` still can't be distributed as a single binary,
because Drake's `find_resource` relies on dynamic loading of the marker
library.