[doc] Add 'serve' for an already-built site
Currently, the 'serve' option first builds the site and then serves it
for preview. Adding a pathway to omit the build and serve whatever
directory the user gives can be useful for developers wanting to
preview a version previously generated elsewhere.

Also, use the `jekyll serve` command instead of a custom TCP server,
as it's more straightforward for this use case.