Extend DRAKE_THROW_UNLESS with value expression arguments
DRAKE_THROW_UNLESS historically took a single parameter: a boolean
expression serving as a throwing condition. if the condition was false,
a message would print out the failed condition.

However, if the condition were expressed in lvalues, the message could
be inscrutable:

    DRAKE_THROW_UNLESS(x > threshold)

would include the following information:

    "Condition 'x > threshold' failed."

with no insight as to what x or threshold are.

Now, up to four value expressions can be passed along with the condition:

    DRAKE_THROW_UNLESS(x > threshold, x, threshold)

Producing (something like):

    "Condition 'x > threshold' failed. x = -inf, threshold = 1e-09."

We exercise the new spelling to enforce requirements on the parameters for
geometry::Shape types.