[lcmtypes] Deprecate fine-grained C++ libraries (#23467)
Deprecate the narrow //lcmtypes:foo cc_library targets and instead use
a single target with all messages for simplicity. The messages don't
change enough to merit the work of fine-grained dependency tracking.

This will also make it easier to port to LCM upstream's new starlark
rules in a future commit.