Change ssao test
CI hates SSAO in the "whole image". The result is so erratically different
from what happens locally, that simply playing with tolerance would be
a bad solution.

However, the standalone test seems quite reasonable. So, we're moving all
of the ssao tests into a smaller, simpler test in hopes that CI won't
choke on it.