[workspace] Use lcm-spy from lcm_internal external (not lcm)
Switch drake-lcm-spy from using `@lcm`'s jar to `@lcm_internal`'s jar,
in preparation for deprecating the `@lcm` external.

Also mark the (unused) python support from `@lcm` with comments in
preparation for deprecating it, and move its regression test from
`//lcm` to `//tools/workspace` so that its removal along with the
deprecation will be more clear.

Between the prior commit that switched over lcm-gen, and this commit
which ditches Java and Python, now the only remaining part of `@lcm`
used by Drake is the C runtime, which will be switched in a future
commit.