[build] Optionally install a static `libdrake.a` (#23298)
Build static and shared copies of `libdrake`, and provide CMake and
Bazel options to decide which one to install. Installing `libdrake`
statically is incompatible with installing Python tools and LCM runtime
support.

Vendor Bazel's `cc_static_library` to build the library, as it was
only de-experimentalized as of Bazel 8.4.0.

For now, `libdrake.a` still can't be distributed as a single binary,
because Drake's `find_resource` relies on dynamic loading of the marker
library.