[pydrake] Rework how docstrings are generated and used (#23431)
Instead of generating docstrings on the fly, we vendor them under
bindings/generated_docstrings. This has two advantages:
- changing a header file doesn't force regeneration of docstrings,
  which has high latency in an edit-test cycle; instead, docstring
  updates are deferred until the linting step where the committed
  docstrings are cross-checked versus the expected docstrings;
- user builds on unsupported platforms don't need to worry about
  providing a compatible libclang.so for header parsing.

In support of this, the mkdoc bazel logic adjusts how header file
filtering happens -- we must tell it an exact physical directory to be
scraped; no more chasing the closure of targets vs target_deps.