[workspace] Add support for Eigen 5.0.0 (#23572)
Upgrade wheel builds to use newer Eigen.

To users -- note that to build Drake with Eigen 5.0.0 requires a patch
to Eigen that has been upstreamed to the 5.0.x branch but not yet tagged
into a release. Users to Drake who use their own version of Eigen 5.0.0
will also need to cherry-pick this patch file, or else wait until Eigen
5.0.1 is released.