Generate venv pins per Python version (#23643)
Tweak the way we generate pins for Python packages to compute separate
pins per each Python version we support. This works around issues with
some packages (numpy, in particular) for which we need a newer package
version with newer Python versions than is supported on the oldest
Python version we support.