[pydrake] Adjust ref count tests for Python 3.14
Python 3.14 introduces new optimizations of reference counting, which caused
some tests to fail as previously written. Adjust tests to pass on a range of
interpreter versions, 3.10 through 3.14.

Also, the effects of the changes can no longer be encapsulated within a
function, so this patch abandons the actual_ref_count() test helper function.