[tools] Disable -Wsign-compare in unit tests
Generally this warning is more noise than it's worth in tests, and
with Bazel 9 it seems like googletest is pickier about sign-compare
in the EXPECT_op macros. Best to just opt-out now.

The opt-out is only for tests (not test_utilities), so we also need to
tweak a few utilities to avoid the forthcoming warnings in Bazel 9.