[render_gl] Remove misleading VTK wrapper comments (#23832)
The call to vtkXDefaultScreen(display()) in my earlier attempt at this
code was ill-formed, leading to the mis-directed blame at VTK macros.

The vtkXDefaultScreen is not actually a function but a global variable
function pointer that gets initialized by the call to display(). Given
that, we must be sure to call display() prior to loading the function
pointer or else the compiler is allowed (but not required?) to load
the pointer prior to evaluating its arguments.