[ad] Special case pow for non-negative exponents of a zero base (#23862)
If base is zero and exponent is non-negative, then the result is now
well-defined per the chain rule (instead of NaN).

This slightly reduces the symmetry of pow with positive vs negative
exponents, but is necessary to support evaluating polynomials because
our default guess for decision variables in mathematical programs is
zero and most solvers cannot handle NaN gradients.