[tools] Automate upgrade check for Doxygen (#23878)
Formalize doxygen_internal's repository file and have it generate
metadata. Add a new type of package upgrades in new_release to check for
upgrades, but not actually do them, to support the workflow for Doxygen.

Also change the verbose option to new_release to write the repository
metadata to a JSON file, instead of dumping to stdout. This is useful
for debugging since Drake's metadata overflows the stdout buffer.