[ad] Store partials in custom vector (#23902)
This is a non-functional refactoring to prepare for future changes
that optimize storage layout, e.g., sparsity, COW, batching, etc.
None of those optimizations are possible with VectorXd storage.

In particular, note that the storage size is now represented by an int
instead of an Eigen::Index. This gives us 4 more unused bytes in the
struct that we'll be able to repurpose for new fields in the future
(e.g., sparsity).